fixed mapper issue and added jpa adapters for owner and restaurant, creating a restaurant use case works completely with a 200 status code

+++ src/main/java/be/kdg/ivanov_kaloyan_prog6_backend/restaurant/adapter/in/request/CreateRestaurantRequest.java
@@ -10,7 +10,7 @@ public record CreateRestaurantRequest(
         AddressDTO address,
         String email,
         String pictureURL,
-        Double defaultPrepTime,
+        Integer defaultPrepTime,
         CuisineType cuisineType,
         OpeningHoursDTO openingHours)
 {
